requirements.txt
beautifulsoup4==4.13.1
boto3==1.36.13
coverage==7.6.10
dj-database-url==2.3.0
django-anymail==12.0
django-bootstrap5==24.3
django-rq==3.0.0
django-storages==1.14.4
django-tempus-dominus==5.1.2.17
Django==5.1.5
easy-thumbnails==2.10
feedparser==6.0.11
gunicorn==23.0.0
markdown==3.7
mock==5.1.0
mutagen==1.47.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
qrcode==8.0
requests==2.32.3
